requirements.txt
pillow==11.0.0
tqdm==4.67.1
